refactor: extract client-construction helpers into _client_setup.py
_helpers.py mixed protocol values (SimpleString, OK), the Database
model, FakeSelector, and glob compilation with ~100 lines of client
kwargs construction used only by the sync/async FakeRedisMixin classes.
Move convert_args_kwargs, build_client_kwds, _get_args_to_warn and
_CONNECTION_POOL_KWARGS to a new _client_setup.py module next to their
only consumers.

Also document what _get_args_to_warn actually does: it walks closure
cells of redis-py's @deprecated_args wrapper to find the args_to_warn
list, since there is no public API for it - the most fragile code in
the package, previously uncommented.

No functional changes; code is moved verbatim.
